feat(program-templates): replace speakersSequence with speaker1-4 + midMeetingInterlude
`speakersSequence` was a single chip that rendered the whole speakers
+ interlude order as a pre-formatted markdown list. Authors couldn't
re-order, restyle, or move the interlude relative to a single
speaker without writing the markdown by hand outside the editor.

Drop the variable entirely and split it into discrete chips:
`speaker1`, `speaker2`, `speaker3`, `speaker4`, `midMeetingInterlude`.
Authors now place each one wherever they want — bullet list, table
row, script cue, etc. — and the print-time renderer plugs in the
real names per meeting.

Also extract the JSON-node builders (`p`, `h`, `ul`, `chip`, …) into
`lexicalNodeBuilders.ts` so the defaults file stays under the LOC
cap with the longer speakers section.

Default conducting + congregation templates now render the speakers
section as a 5-row bulleted list (Speaker 1 / Speaker 2 / Mid-meeting
interlude / Speaker 3 / Speaker 4) so wards with fewer speakers see
empty spots they can delete cleanly.
